requirements.txt
Flask==2.3.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
requests==2.31.0
torch==2.0.1
transformers==4.33.2